vinzapai-backend/requirements.txt
django>=5.2,<6.0
djangorestframework>=3.15
django-jazzmin>=3.0
django-cors-headers>=4.3
whitenoise>=6.7
waitress>=3.0
Pillow>=10.0
pywebview>=5.0
python-decouple>=3.8
